0_test/layout-simulator-service/requirements.txt
fastapi
uvicorn
Pillow
ultralytics
